联盟优惠项目

+++ src/main/java/com/gt/union/controller/basic/UnionMemberPreferentialServiceController.java
@@ -1,9 +1,13 @@
 package com.gt.union.controller.basic;
 
+import com.baomidou.mybatisplus.plugins.Page;
 import com.gt.union.common.annotation.SysLogAnnotation;
+import com.gt.union.common.constant.CommonConstant;
 import com.gt.union.common.constant.ExceptionConstant;
 import com.gt.union.common.exception.BaseException;
+import com.gt.union.common.exception.BusinessException;
 import com.gt.union.common.response.GTJsonResult;
+import com.gt.union.common.util.CommonUtil;
 import com.gt.union.common.util.SessionUtils;
 import com.gt.union.entity.common.BusUser;
 import com.gt.union.service.basic.IUnionMemberPreferentialServiceService;
@@ -31,11 +35,33 @@ public class UnionMemberPreferentialServiceController {
 	private static final String DELETE = "UnionMemberPreferentialServiceController.delete()";
 	private static final String VERIFY = "UnionMemberPreferentialServiceController.verify()";
 	private static final String ADD_SERVICE_VERIFY = "UnionMemberPreferentialServiceController.addServiceVerify()";
+	private static final String LIST_MY_BY_UNIONID = "UnionMemberPreferentialServiceController.listMyByUnionId()";
 	private Logger logger = LoggerFactory.getLogger(UnionMemberPreferentialServiceController.class);
 
 	@Autowired
 	private IUnionMemberPreferentialServiceService unionMemberPreferentialServiceService;
 
+
+	@ApiOperation(value = "查询我的优惠项目", produces = "application/json;charset=UTF-8")
+	@RequestMapping(value = "/unionId/{unionId}/list", method = RequestMethod.GET, produces = "application/json;charset=UTF-8")
+	public String listMyByUnionId(Page page, HttpServletRequest request, @ApiParam(name = "unionId", value = "联盟id", required = true) @PathVariable(name = "unionId", required = true) Integer unionId
+			) {
+		try {
+			BusUser busUser = SessionUtils.getLoginUser(request);
+			Integer busId = busUser.getId();
+			if(CommonUtil.isNotEmpty(busUser.getPid()) && busUser.getPid() != 0){
+				busId = busUser.getPid();
+			}
+			Page result = this.unionMemberPreferentialServiceService.listMyByUnionId(page,unionId, busId);
+			return GTJsonResult.instanceSuccessMsg(result).toString();
+		} catch (BaseException e) {
+			logger.error("", e);
+			return GTJsonResult.instanceErrorMsg(e.getErrorLocation(), e.getErrorCausedBy(), e.getErrorMsg()).toString();
+		} catch (Exception e) {
+			logger.error("", e);
+			return GTJsonResult.instanceErrorMsg(LIST_MY_BY_UNIONID, e.getMessage(), ExceptionConstant.OPERATE_FAIL).toString();
+		}
+	}
 
 	@ApiOperation(value = "保存我的优惠项目", produces = "application/json;charset=UTF-8")
 	@SysLogAnnotation(op_function = "2", description = "保存我的优惠项目")
@@ -44,6 +70,9 @@ public class UnionMemberPreferentialServiceController {
 			, @ApiParam(name = "serviceName", value = "项目名称", required = true) @RequestParam(name = "serviceName", required = true) String serviceName) {
 		try {
 			BusUser busUser = SessionUtils.getLoginUser(request);
+			if(CommonUtil.isNotEmpty(busUser.getPid()) && busUser.getPid() != 0){
+				throw new BusinessException(SAVE, "", CommonConstant.UNION_BUS_PARENT_MSG);
+			}
 			this.unionMemberPreferentialServiceService.save(unionId, busUser.getId(), serviceName);
 			return GTJsonResult.instanceSuccessMsg().toString();
 		} catch (BaseException e) {
@@ -61,6 +90,10 @@ public class UnionMemberPreferentialServiceController {
 	public String listMyByUnionId(HttpServletRequest request, @ApiParam(name = "unionId", value = "联盟id", required = true) @PathVariable(name = "unionId", required = true) Integer unionId
 			,@ApiParam(name = "ids", value = "项目ids", required = true) @RequestParam(name = "ids", required = true) String ids) {
 		try {
+			BusUser busUser = SessionUtils.getLoginUser(request);
+			if(CommonUtil.isNotEmpty(busUser.getPid()) && busUser.getPid() != 0){
+				throw new BusinessException(SAVE, "", CommonConstant.UNION_BUS_PARENT_MSG);
+			}
 			unionMemberPreferentialServiceService.delete(unionId, ids);
 			return GTJsonResult.instanceSuccessMsg().toString();
 		} catch (BaseException e) {
@@ -74,10 +107,14 @@ public class UnionMemberPreferentialServiceController {
 
 	@ApiOperation(value = "优惠项目提交审核", produces = "application/json;charset=UTF-8")
 	@SysLogAnnotation(op_function = "3", description = "优惠项目提交审核")
-	@RequestMapping(value = "/{id}/unionId/{unionId}", method = RequestMethod.PATCH, produces = "application/json;charset=UTF-8")
+	@RequestMapping(value = "/{id}/unionId/{unionId}", method = RequestMethod.POST, produces = "application/json;charset=UTF-8")
 	public String addServiceVerify(HttpServletRequest request, @ApiParam(name = "unionId", value = "联盟id", required = true) @PathVariable(name = "unionId", required = true) Integer unionId
 			,@ApiParam(name = "id", value = "项目id", required = true) @PathVariable(name = "id", required = true) Integer id) {
 		try {
+			BusUser busUser = SessionUtils.getLoginUser(request);
+			if(CommonUtil.isNotEmpty(busUser.getPid()) && busUser.getPid() != 0){
+				throw new BusinessException(SAVE, "", CommonConstant.UNION_BUS_PARENT_MSG);
+			}
 			unionMemberPreferentialServiceService.addServiceVerify(unionId, id);
 			return GTJsonResult.instanceSuccessMsg().toString();
 		} catch (BaseException e) {
